# api/openapi.yaml
openapi: 3.0.3
info:
  title: Symposium API
  version: 1.0.0
  description: |
    Read a Symposium community record, follow it as it changes, and publish through the gate.

    The Symposium API speaks in the specification's model: Artifacts, Objects, properties,
    relationships, Members and addresses (`spec/symposium_specification.md`, profiled by
    `tools/CANONICAL.md`). It is served by the Symposium Data server beside its file-level
    `/v1` API, and reads the community record held there, so every application sees the one
    record under the server's own access rules.

    **The record is read-only here.** Artifacts are immutable. A correction is a new Artifact
    that supersedes the old one, published through `POST /{community}/submissions` and decided
    by the gate. The API offers no update and no delete of a recorded Artifact.

    **Canonical URLs.** Every resource carries a `url`: an absolute URL on the data server that
    names it whichever host asks. An address maps to a URL by a fixed rule (see
    `GET /{community}/resolve` and `api/DESIGN.md`).

    **Freshness.** Every read response carries `position`: the record's cursor and the
    `created` of the newest Artifact the response reflects.

    **Paging.** Every listing whose size grows with the record takes `cursor` and `limit`
    (1 to 1000, default 100) and answers `next`, the cursor of the following page, or `null`
    on the last page. Each listing states its order.

...

...

...

...

...
  license:
    name: MIT
...
servers:
...
security:
  - apiKey: []
paths:
...
  /{community}/submissions/check:
    post:
      operationId: checkSubmission
      tags: [publishing]
      summary: Validate an Artifact and submit nothing
      description: |
        The checks the gate makes, against the record as it stands, with a provisional
        `created` from the server's clock: the naming rule, the 250 KB limit on embedded
        payload, and the validator. Nothing is stored. A pass means the gate accepts the
        Artifact if the record has not moved. This is `/symposium validate` and
        `publish --check`.

        Roles: member, admin.
      x-roles: [member, admin]
      x-bounded: the findings on the one Artifact submitted
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
      requestBody:
        $ref: '#/components/requestBodies/Submission'
      responses:
        '200':
          description: The validator's verdict. `ok` is false when any finding is a FAIL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionCheck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
  /{community}/submissions:
    get:
      operationId: listSubmissions
      tags: [publishing]
      summary: Submissions and their decisions
      description: |
        Submissions in `inbox` with the gate's decision on each. A `member` key sees its own; an
        `admin` key sees every one. Ordered by submission time ascending, the order the gate
        reads them in.

        Roles: member, admin.
      x-roles: [member, admin]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          description: Only submissions in this state.
          schema:
            $ref: '#/components/schemas/SubmissionStatus'
      responses:
        '200':
          description: One page of submissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: submitArtifact
      tags: [publishing]
      summary: Publish one Artifact through the gate
      description: |
        Validate one Artifact as `checkSubmission` does, then store it in `inbox` as a
        submission attributed to the key's Member, for the gate to decide. The Artifact is not
        in the record until the gate accepts it; the gate stamps `created` then. One Artifact
        per call, as `/symposium publish` takes one: publication is serial.

        `published_by` must be the key's Member, `@<username>`, and `name` must carry its
        prefix. A submission the validator fails is refused with 422 and its findings, and
        nothing is stored. This is `/symposium publish`.

        Roles: member, admin.
      x-roles: [member, admin]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
      requestBody:
        $ref: '#/components/requestBodies/Submission'
      responses:
        '201':
          description: The submission, pending the gate's decision.
          headers:
            Location:
              description: The submission's canonical URL.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /{community}/submissions/{submission}:
    get:
      operationId: getSubmission
      tags: [publishing]
      summary: One submission and its decision
      description: |
        One submission, its state, and the gate's decision on it: the record Artifact when
        accepted, the gate's reply with its failures when rejected. A `member` key reads only
        its own.

        Roles: member, admin.
      x-roles: [member, admin]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/submission'
      responses:
        '200':
          description: The submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sak_<43 base64url characters>
      description: |
        An API key the admin issued, sent as `Authorization: Bearer sak_…`. The `sak_` prefix
        tells it apart from the data server's Ed25519 tokens (JWTs) and read keys (`sdr_…`),
        which share the same header. The server matches a key by its SHA-256 digest; it
        decrypts nothing to authenticate.
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The server admin's Ed25519 token from `POST /v1/admin/token`, the one the skill already
        holds. The API-key operations accept it alone, so an API key, whatever its scope,
        reads and makes no other key.
  parameters:
    community:
      name: community
      in: path
      required: true
      description: The community's name, matched ignoring case.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_]{1,20}$'
...
    submission:
      name: submission
      in: path
      required: true
      description: The submission's id, the id of its file in `inbox`.
      schema:
        type: string
        format: uuid
...
    cursor:
      name: cursor
      in: query
      description: The `next` of the previous page. Absent for the first page.
      schema:
        type: string
        maxLength: 256
        pattern: '^[A-Za-z0-9_-]+$'
    limit:
      name: limit
      in: query
      description: The most items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
...
  requestBodies:
    Submission:
      required: true
      description: One Artifact in canonical JSON, with `created` null.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SubmittedArtifact'
  responses:
    BadRequest:
      description: A parameter or the body is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: The credential is absent, unknown, expired or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid and its roles or scope do not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The community or the resource is unknown.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: |
        The request conflicts with the state: a name already in the record or already
        submitted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooLarge:
      description: The body is over 8 MiB, or over the Member's quota.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationFailed:
      description: The validator failed the Artifact. `findings` lists every finding.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: |
        The one error body. `detail` is a sentence for a person, as the data server's `/v1`
        writes it; `code` is stable for a program.
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - too_large
            - validation_failed
            - unprocessable
            - not_operational
            - too_many_streams
        findings:
          type: array
          description: The validator's findings, on `validation_failed` only.
          items:
            $ref: '#/components/schemas/Finding'
...
    Handle:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$
    Name:
      type: string
      description: A name in the specification's sense. It holds no `.` or `#` and starts with no `@`.
      pattern: ^[^@.#][^.#]*$
    Address:
      type: string
      description: |
        An address (specification §1.8): `@` and dot-separated segments naming a Member, an
        Artifact, an Object or a property, then optionally `#` and a schema reference.
      pattern: ^@[^.#]+
      example: '@agent_lyra_data_overexpression_screen_v1.values#csv.row=BST2&col=Normalized infection'
    Cursor:
      type: string
      maxLength: 256
      pattern: ^[A-Za-z0-9_-]+$
      description: |
        An opaque position in an ordered listing, base64url. It stays valid for the life of the
        record.
    Position:
      type: object
      description: The state of the record a read response reflects.
      required: [cursor, seq, created, as_of]
      properties:
        cursor:
          $ref: '#/components/schemas/Cursor'
        seq:
          type: integer
          minimum: 0
          description: The `record` collection's sequence number at the newest Artifact reflected.
        created:
          type: string
          nullable: true
          format: date-time
          description: The `created` of the newest Artifact reflected; null for an empty record.
        as_of:
          type: string
          format: date-time
          description: When the server read the record for this response.
...
    ArtifactHeader:
      type: object
      description: |
        An Artifact's properties (specification §1.5 and §2, profile §1 and §2.1). Members may
        add properties of their own. Each extra property's value is a `PropertyValue`.
      required: [name, type, created, published_by, specification_version]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        type:
          type: string
          description: Argument, Data, ScientificPublication, Analysis, Model, NonGroundable, Message, ResearchGoal, or a type of the Member's own.
        created:
          type: string
          format: date-time
          nullable: true
          description: |
            Stamped by the gate at acceptance, so every record Artifact carries it. A submission
            sends null, and `checkSubmission` and `submitArtifact` refuse any other value.
        published_by:
          $ref: '#/components/schemas/Address'
        specification_version:
          type: string
        title:
          type: string
        description:
          type: string
        text:
          type: string
        authors:
          type: array
          items:
            type: string
        supersedes:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        supersedes_rationale:
          type: string
        import_method:
          type: string
        produced_by:
          $ref: '#/components/schemas/Address'
        groundable:
          type: boolean
        serves_goals:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        primary_assertion:
          type: string
          description: Argument. The name of its primary Assertion.
        verdict:
          type: string
          description: Argument.
        rationale:
          type: string
          description: Argument.
        purpose:
          type: string
          description: Argument.
        genre:
          type: string
          description: Argument, profile. `finding` or `assessment`.
        extracted_from:
          $ref: '#/components/schemas/Address'
        extraction_method:
          type: string
        procedure:
          type: string
          description: Analysis.
        inputs:
          type: array
          description: Analysis.
          items:
            $ref: '#/components/schemas/Address'
        modeling_choices:
          type: string
          description: Model.
        recipients:
          type: array
          description: Message.
          items:
            $ref: '#/components/schemas/Address'
      additionalProperties: true
    RecordObject:
      type: object
      description: |
        An Object (specification §1.6). Assertion, Ground, Assumption and Content carry the
        properties the specification requires of them; any other type is the author's.
        Each extra property's value is a `PropertyValue`.
      required: [name, type]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        type:
          type: string
        title:
          type: string
        description:
          type: string
        text:
          type: string
        claim:
          type: string
          description: Assertion.
        scope:
          type: string
          description: Assertion.
        citation:
          $ref: '#/components/schemas/Address'
        rationale:
          type: string
          description: Ground, Assumption.
        criterion:
          type: string
          description: Ground. Its presence marks the Ground as a test.
        addressing_method:
          type: string
          description: Content.
        groundable:
          type: boolean
          description: Content.
        location:
          type: string
          description: Content.
        access_method:
          type: string
          description: Content.
        sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          description: Content, profile.
      additionalProperties: true
    Relationship:
      type: object
      description: |
        A directed relationship between two Objects of one Artifact (specification §1.7).
        Each extra property's value is a `PropertyValue`.
      required: [rel, source, target]
      properties:
        rel:
          type: string
        source:
          $ref: '#/components/schemas/Name'
        target:
          $ref: '#/components/schemas/Name'
      additionalProperties: true
...
    SubmittedArtifact:
      type: object
      description: One Artifact in canonical JSON as its author submits it.
      required: [artifact, objects, relationships]
      properties:
        artifact:
          $ref: '#/components/schemas/ArtifactHeader'
        objects:
          type: array
          items:
            $ref: '#/components/schemas/RecordObject'
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
    Page:
      type: object
      description: The fields every paged listing carries.
      required: [next, position]
      properties:
        next:
          description: The cursor of the following page; null on the last page.
          type: string
          nullable: true
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
        position:
          $ref: '#/components/schemas/Position'
...
    FindingLevel:
      type: string
      enum: [FAIL, REVIEW]
    Finding:
      type: object
      description: One finding of the validator, as `tools/validate.py` writes it.
      required: [check, level, msg]
      properties:
        check:
          type: string
          description: STRUCT, TYPE, ADDRESS, GROUND, CITATION, GOAL, INDEPENDENCE, NAMING, ORDER, SIZE, SUPERSEDED, UNIQUE.
        level:
          $ref: '#/components/schemas/FindingLevel'
        msg:
          type: string
        object:
          description: The Object the finding names, where it names one.
          allOf:
            - $ref: '#/components/schemas/Name'
...
    SubmissionCheck:
      type: object
      required: [ok, findings, position]
      properties:
        ok:
          type: boolean
        findings:
          type: array
          items:
            $ref: '#/components/schemas/Finding'
        position:
          $ref: '#/components/schemas/Position'
    SubmissionStatus:
      type: string
      description: |
        `pending` until the gate decides. `skipped` when the gate's own rule sets it aside: its
        inbox name is the artifact's name with the wrong prefix, or its submitter is off the
        roster. Then `accepted` or `rejected`. `api/DESIGN.md` §6 says where each comes from.
      enum: [pending, skipped, accepted, rejected]
    SubmissionResource:
      type: object
      required: [id, url, name, submitted_by, submitted, status, citation, position]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          $ref: '#/components/schemas/Name'
        type:
          type: string
        submitted_by:
          $ref: '#/components/schemas/Handle'
        submitted:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        citation:
          type: string
          description: The submission's file version on the data server, `symposium-data:<uuid>@v1`.
        decided:
          type: string
          nullable: true
          format: date-time
        artifact_url:
          type: string
          format: uri
          description: Accepted. The record Artifact.
        reply:
          description: Rejected. The gate's reply.
          allOf:
            - $ref: '#/components/schemas/Reply'
        position:
          $ref: '#/components/schemas/Position'
    SubmissionPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SubmissionResource'
    Reply:
      type: object
      description: The gate's reply to a rejected submission, as it sits in `inbox`.
      required: [sender, sent, text, failures]
      properties:
        sender:
          $ref: '#/components/schemas/Handle'
        sent:
          type: string
          format: date-time
        text:
          type: string
          description: The reply as the gate wrote it, `REJECTED` and one line per failure.
        failures:
          type: array
          description: One line per failure the validator found.
          items:
            type: string
...