# data-server/service/symposium_api/generated/routers/publishing.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["publishing"])


@router.get(
    "/{community}/submissions",
    response_model=SubmissionPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listSubmissions",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_submissions(
    community: str,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    status: Optional[SubmissionStatus] = None,
    caller: Caller = Depends(authorize("listSubmissions")),
    service: Service = Depends(service_provider),
) -> SubmissionPage:
    """Submissions and their decisions"""
    return service.list_submissions(
        caller, community=community, cursor=cursor, limit=limit, status=status
    )


@router.post(
    "/{community}/submissions",
    response_model=SubmissionResource,
    response_model_exclude_unset=True,
    status_code=201,
    operation_id="submitArtifact",
    responses={
        "201": {
            "model": SubmissionResource,
        },
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
        "409": {
            "model": Error,
        },
        "413": {
            "model": Error,
        },
        "422": {
            "model": Error,
        },
    },
)
def submit_artifact(
    community: str,
    body: SubmittedArtifact = ...,
    caller: Caller = Depends(authorize("submitArtifact")),
    service: Service = Depends(service_provider),
) -> SubmissionResource:
    """Publish one Artifact through the gate"""
    return service.submit_artifact(caller, community=community, body=body)


@router.post(
    "/{community}/submissions/check",
    response_model=SubmissionCheck,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="checkSubmission",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
        "413": {
            "model": Error,
        },
    },
)
def check_submission(
    community: str,
    body: SubmittedArtifact = ...,
    caller: Caller = Depends(authorize("checkSubmission")),
    service: Service = Depends(service_provider),
) -> SubmissionCheck:
    """Validate an Artifact and submit nothing"""
    return service.check_submission(caller, community=community, body=body)


@router.get(
    "/{community}/submissions/{submission}",
    response_model=SubmissionResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getSubmission",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_submission(
    community: str,
    submission: UUID = ...,
    caller: Caller = Depends(authorize("getSubmission")),
    service: Service = Depends(service_provider),
) -> SubmissionResource:
    """One submission and its decision"""
    return service.get_submission(caller, community=community, submission=submission)